refactor(web): align chat store with CONVENTIONS.md and STYLE_GUIDE.md
- Rename chat-store.ts → use-chat-store.ts (use-{domain}-store.ts convention)
- Rename ChatContextValue → ChatStore, ChatStateValue → ChatState,
  ChatActionsValue → ChatActions (remove Context-era naming)
- Clean up module docstring to describe current code, not migration history
- Update imports in chat-page.tsx and test file
