+++ apps/web/src/domains/chat/use-chat-store.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeEach } from "bun:test";
 
-import { useChatStore } from "@/domains/chat/chat-store.js";
+import { useChatStore } from "@/domains/chat/use-chat-store.js";
 
 beforeEach(() => {
   useChatStore.setState({
